+++ src/components/index.ts
@@ -5,3 +5,4 @@ export { Header } from './Header'
 export { LabelledContent } from './LabelledContent'
 export { OnboardingContainer } from './OnboardingContainer'
 export { Progress } from './Progress'
+export { Tooltip } from './Tooltip'
